Login and Customer update
Fixed login so that it checks username and password with the database instead of just comparing to a string value. 

Built out the delete customer method and the insert customer method. Both still need to be altered to use the logged in users name once variable passing is implemented between controllers.